requirements.txt
django-oscar==3.2.4
sorl-thumbnail==12.11.0
pysolr==3.10.0
Pillow>=8.0.0
